Added math
Currently, adding, subtracting, and multiplying are supported.

The parser works by recursively going through the order of operations
looking for an addition or subtraction, then within that a
multiplication, then within that just a number, but that can start the
whole loop.

The syntax tree then makes the code gen quite easy