Functions code gen
It works in a similar way, the generator splits the assembly into two
parts, the main and afterwards for function definitions. It then calls
itself recursively for scopes. It does not yet support arguments, will
be added soon. I also added some commenting to the code to make it a bit
clearer.